RestAPI_Task/requirements.txt
aiofiles==24.1.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.19.1
pillow==10.3.0
python-multipart==0.0.6
